backend/requirements.txt
# ============================================================================
# Farm Management System - Backend Dependencies
# ============================================================================
# Version: 1.0.0
# Created: 2025-11-17
#
# Changelog:
# ----------
# v1.0.0 (2025-11-17):
#   - Initial dependencies for FastAPI backend
#   - Async PostgreSQL support
#   - JWT authentication
#   - Supabase integration
#   - Excel export capabilities
#   - Testing framework
# ============================================================================

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database - Async PostgreSQL
asyncpg==0.29.0
databases[postgresql]==0.8.0
sqlalchemy==2.0.23

# Supabase
supabase==2.3.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Data Processing & Export
pandas==2.1.3
xlsxwriter==3.1.9
openpyxl==3.1.2

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
ruff==0.1.8

# CORS & Middleware
python-cors==1.0.0
